fix(seyond): suppress unused variable and type comparison warnings
Remove unused variables and add explicit cast to fix compiler warnings in seyond_decoder.

+++ nebula_decoders/include/nebula_decoders/nebula_decoders_seyond/decoders/seyond_decoder.hpp
@@ -261,7 +261,7 @@ public:
     int h_index = h_angle >> kTableShift_;
     // avoid index out-of-bound
     h_index = h_index & (kHTableSize_ - 1);
-    if (h_index > kHTableSize_ - 2) {
+    if (h_index > static_cast<int32_t>(kHTableSize_) - 2) {
       h_index = kHTableSize_ - 2;
     }
